mongo/requirements.txt
pymongo
azure-storage-blob
python-dotenv
